Clean up services and file structure
