+++ src/components/posts/AllPosts.jsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from "react";
-import { fetchPosts } from "../../managers/PostManager.js";
+import { fetchPosts } from "../../services/postService.js";
 import { Link } from "react-router-dom";
 import "./AllPosts.css"
