+++ src/services/postService.js
@@ -1,3 +1,13 @@
+export const fetchPosts = () => {
+    return fetch("http://localhost:8000/posts?_expand=category", {
+      method: "GET",
+      headers: {
+        "Content-Type": "application/json",
+        "Accept": "application/json",
+      },
+    }).then((res) => res.json());
+  };
+
 export const getAllPosts = async () => {
     return fetch("http://localhost:8000/posts").then((res) =>
         res.json()
